app/requirements.txt
flask==2.3.2
python-dotenv==1.0.0
assemblyai==0.9.1
requests==2.31.0
pymongo==4.5.0
openai==0.28.0